more tests to correct file
